+++ tools/decode-status.py
@@ -182,8 +182,10 @@ a_setoption = [[
     "(Zigbee) Hide bridge topic from zigbee topic (use with SetOption89) (1)",
     "(DS18x20) Enable arithmetic mean over teleperiod for JSON temperature (1)",
     "(Wifi) Keep wifi in no-sleep mode, prevents some occasional unresponsiveness",
-    "","",
-    "","","","",
+    "(Web) Allow access without referer check",
+    "(Energy) Show phase information",
+    "(Debug) Show heap with logging timestamp",
+    "","","",
     "","","","",
     "","","","",
     "","","",""
@@ -256,7 +258,7 @@ a_features = [[
     "USE_MPU_ACCEL","USE_TFMINIPLUS","USE_CSE7761","USE_BERRY",
     "USE_BM8563","USE_ENERGY_DUMMY","USE_AM2320","USE_T67XX",
     "USE_MCP2515","USE_TASMESH","USE_WIFI_RANGE_EXTENDER","USE_INFLUXDB",
-    "USE_HRG15","USE_VINDRIKTNING","USE_SCD40","",
+    "USE_HRG15","USE_VINDRIKTNING","USE_SCD40","USE_HM330X",
     "","","","",
     "","","","",
     "","","","",
@@ -288,7 +290,7 @@ else:
         obj = json.load(fp)
 
 def StartDecode():
-    print ("\n*** decode-status.py v20210901 by Theo Arends and Jacek Ziolkowski ***")
+    print ("\n*** decode-status.py v20211008 by Theo Arends and Jacek Ziolkowski ***")
 
 #    print("Decoding\n{}".format(obj))